+++ src/app/system.rs
@@ -1,9 +1,12 @@
-use crate::app::ui::{key, key_val, val};
+use crate::app::ui::{icon, key, key_val, val};
 use crate::app::App;
+use crate::app::{containers, images};
 use crate::event::SystemInspectInfo;
 
 use docker_api::api::DataUsage;
 use egui::{CollapsingHeader, Grid};
+
+const MAX_ITEM_COUNT: usize = 10;
 
 #[derive(Clone, Copy, PartialEq, Debug)]
 pub enum CentralView {
@@ -22,6 +25,9 @@ pub struct SystemTab {
     pub system_info: Option<Box<SystemInspectInfo>>,
     pub data_usage: Option<Box<DataUsage>>,
     pub central_view: CentralView,
+    pub display_all_containers: bool,
+    pub display_all_images: bool,
+    pub display_all_cache: bool,
 }
 
 impl App {
@@ -48,7 +54,256 @@ impl App {
     }
 
     fn system_data_usage(&mut self, ui: &mut egui::Ui) {
-        ui.label("TODO!");
+        ui.allocate_space((f32::INFINITY, 0.).into());
+        self.containers_data_usage(ui);
+        ui.add_space(10.);
+        self.images_data_usage(ui);
+        ui.add_space(10.);
+        self.build_cache_usage(ui);
+    }
+
+    fn containers_data_usage(&mut self, ui: &mut egui::Ui) {
+        CollapsingHeader::new("containers")
+            .default_open(false)
+            .show(ui, |ui| {
+                if let Some(usage) = self.system.data_usage.as_ref() {
+                    let (count, total_root, total_rw) =
+                        usage.containers.iter().fold((0, 0, 0), |mut acc, c| {
+                            acc.0 += 1;
+                            acc.1 += c.size_root_fs.unwrap_or_default() as u64;
+                            acc.2 += c.size_rw.unwrap_or_default() as u64;
+                            acc
+                        });
+                    ui.checkbox(&mut self.system.display_all_containers, "Display all");
+                    ui.add_space(10.);
+                    Grid::new("total_size_grid").show(ui, |ui| {
+                        key!(ui, "Total size:");
+                        Grid::new("size_details_grid").show(ui, |ui| {
+                            ui.label("Root FS");
+                            ui.label("RW");
+                            ui.end_row();
+                            val!(ui, crate::conv_b(total_root));
+                            val!(ui, crate::conv_b(total_rw));
+                        });
+                    });
+                    Grid::new("containers_grid")
+                        .spacing((20., 10.))
+                        .min_col_width(50.)
+                        .show(ui, |ui| {
+                            key!(ui, "ID");
+                            key!(ui, "Created");
+                            key!(ui, "Image");
+                            key!(ui, "Command");
+                            key!(ui, "Size Total");
+                            key!(ui, "Size RW");
+                            ui.end_row();
+                            let mut containers = usage.containers.iter().collect::<Vec<_>>();
+                            containers.sort_by(|a, b| b.size_root_fs.cmp(&a.size_root_fs));
+
+                            fn container_stats(
+                                ui: &mut egui::Ui,
+                                container: &docker_api::api::ContainerSummary,
+                            ) {
+                                let name = if let Some(first) = container.names.first() {
+                                    first.trim_start_matches('/')
+                                } else {
+                                    &container.id[0..12]
+                                };
+                                let color = containers::color_for_state(&container.state);
+                                let icon = containers::state_icon(color);
+                                ui.scope(|ui| {
+                                    ui.add(icon);
+                                    val!(ui, name);
+                                });
+                                let naive =
+                                    chrono::NaiveDateTime::from_timestamp(container.created, 0);
+                                let datetime: chrono::DateTime<chrono::Utc> =
+                                    chrono::DateTime::from_utc(naive, chrono::Utc);
+                                let command = if container.command.len() > 32 {
+                                    &container.command[..32]
+                                } else {
+                                    &container.command[..]
+                                };
+                                val!(ui, datetime.to_rfc2822());
+                                val!(ui, images::trim_id(&container.image));
+                                val!(ui, command);
+                                val!(
+                                    ui,
+                                    crate::conv_b(container.size_root_fs.unwrap_or_default() as u64)
+                                );
+                                val!(
+                                    ui,
+                                    crate::conv_b(container.size_rw.unwrap_or_default() as u64)
+                                );
+                            }
+
+                            if self.system.display_all_containers {
+                                for container in containers {
+                                    container_stats(ui, container);
+                                    ui.end_row();
+                                }
+                            } else {
+                                for container in containers.iter().take(MAX_ITEM_COUNT) {
+                                    container_stats(ui, container);
+                                    ui.end_row();
+                                }
+                            }
+                        });
+                    if !self.system.display_all_containers && count > MAX_ITEM_COUNT {
+                        ui.add(egui::Label::new("More to load...").weak());
+                    }
+                }
+            });
+    }
+
+    fn images_data_usage(&mut self, ui: &mut egui::Ui) {
+        CollapsingHeader::new("images")
+            .default_open(false)
+            .show(ui, |ui| {
+                if let Some(usage) = self.system.data_usage.as_ref() {
+                    let (count, total, total_shared, total_virtual) =
+                        usage.images.iter().fold((0, 0, 0, 0), |mut acc, i| {
+                            acc.0 += 1;
+                            acc.1 += i.size;
+                            acc.2 += i.shared_size;
+                            acc.3 += i.virtual_size;
+                            acc
+                        });
+                    ui.checkbox(&mut self.system.display_all_images, "Display all");
+                    ui.add_space(10.);
+                    Grid::new("total_images_size_grid").show(ui, |ui| {
+                        key!(ui, "Total size:");
+                        Grid::new("images_size_details_grid").show(ui, |ui| {
+                            key!(ui, "Total");
+                            key!(ui, "Shared");
+                            key!(ui, "Virtual");
+                            ui.end_row();
+                            val!(ui, crate::conv_b(total as u64));
+                            val!(ui, crate::conv_b(total_shared as u64));
+                            val!(ui, crate::conv_b(total_virtual as u64));
+                        });
+                    });
+                    Grid::new("images_grid")
+                        .spacing((20., 10.))
+                        .min_col_width(50.)
+                        .show(ui, |ui| {
+                            key!(ui, "ID");
+                            key!(ui, "Created");
+                            key!(ui, "Containers");
+                            key!(ui, "Size");
+                            key!(ui, "Shared size");
+                            key!(ui, "Virtual size");
+                            ui.end_row();
+                            let mut images = usage.images.iter().collect::<Vec<_>>();
+                            images.sort_by(|a, b| b.size.cmp(&a.size));
+
+                            fn image_stats(
+                                ui: &mut egui::Ui,
+                                image: &docker_api::api::ImageSummary,
+                            ) {
+                                let name = if let Some(first) = image.repo_tags.first() {
+                                    first
+                                } else {
+                                    images::trim_id(&image.id)
+                                };
+                                let naive =
+                                    chrono::NaiveDateTime::from_timestamp(image.created as i64, 0);
+                                let datetime: chrono::DateTime<chrono::Utc> =
+                                    chrono::DateTime::from_utc(naive, chrono::Utc);
+
+                                ui.scope(|ui| {
+                                    ui.add(images::icon());
+                                    val!(ui, name);
+                                });
+                                val!(ui, datetime.to_rfc2822());
+                                val!(ui, image.containers);
+                                val!(ui, crate::conv_b(image.size as u64));
+                                val!(ui, crate::conv_b(image.shared_size as u64));
+                                val!(ui, crate::conv_b(image.virtual_size as u64));
+                            }
+
+                            if self.system.display_all_images {
+                                for image in images {
+                                    image_stats(ui, image);
+                                    ui.end_row();
+                                }
+                            } else {
+                                for image in images.iter().take(MAX_ITEM_COUNT) {
+                                    image_stats(ui, image);
+                                    ui.end_row();
+                                }
+                            };
+                        });
+                    if !self.system.display_all_images && count > MAX_ITEM_COUNT {
+                        ui.add(egui::Label::new("More to load...").weak());
+                    }
+                }
+            });
+    }
+
+    fn build_cache_usage(&mut self, ui: &mut egui::Ui) {
+        if let Some(usage) = self.system.data_usage.as_ref() {
+            CollapsingHeader::new("build cache")
+                .default_open(false)
+                .show(ui, |ui| {
+                    if let Some(cache) = &usage.build_cache {
+                        let (count, total) = usage.images.iter().fold((0, 0), |mut acc, i| {
+                            acc.0 += 1;
+                            acc.1 += i.size;
+                            acc
+                        });
+                        ui.checkbox(&mut self.system.display_all_cache, "Display all");
+                        ui.add_space(10.);
+                        key_val!(ui, "Total size:", crate::conv_b(total as u64));
+                        Grid::new("cache_grid")
+                            .spacing((20., 10.))
+                            .min_col_width(50.)
+                            .show(ui, |ui| {
+                                key!(ui, "ID");
+                                key!(ui, "Type");
+                                key!(ui, "Created");
+                                key!(ui, "In use");
+                                key!(ui, "Shared");
+                                key!(ui, "Usage count");
+                                key!(ui, "Size");
+                                ui.end_row();
+                                let mut images = cache.iter().collect::<Vec<_>>();
+                                images.sort_by(|a, b| b.size.cmp(&a.size));
+
+                                fn cache_stats(
+                                    ui: &mut egui::Ui,
+                                    cache: &docker_api::api::BuildCache,
+                                ) {
+                                    ui.scope(|ui| {
+                                        ui.add(egui::Label::new(icon::DISK).heading().strong());
+                                        val!(ui, images::trim_id(&cache.id));
+                                    });
+                                    val!(ui, &cache.type_);
+                                    val!(ui, cache.created_at.to_rfc2822());
+                                    val!(ui, cache.in_use);
+                                    val!(ui, cache.shared);
+                                    val!(ui, cache.usage_count);
+                                    val!(ui, crate::conv_b(cache.size as u64));
+                                }
+
+                                if self.system.display_all_cache {
+                                    for cache in images {
+                                        cache_stats(ui, cache);
+                                        ui.end_row();
+                                    }
+                                } else {
+                                    for cache in images.iter().take(MAX_ITEM_COUNT) {
+                                        cache_stats(ui, cache);
+                                        ui.end_row();
+                                    }
+                                };
+                            });
+                        if !self.system.display_all_cache && count > MAX_ITEM_COUNT {
+                            ui.add(egui::Label::new("More to load...").weak());
+                        }
+                    }
+                });
+        }
     }
 
     fn system_details(&mut self, ui: &mut egui::Ui) {